+++ app/playlist_ui/trackinfodialog.h
@@ -2,31 +2,45 @@
 #define TRACKINFODIALOG_H
 
 #include "track.h"
+#include "playlist/playlist.h"
 
 #include <QDialog>
 #include <QStandardItemModel>
+#include <memory>
 
 namespace Ui {
   class TrackInfoDialog;
 }
 
+class TagEditorDialog;
+
 class TrackInfoDialog : public QDialog {
   Q_OBJECT
 
 public:
-  explicit TrackInfoDialog(const Track &track, QWidget *parent = nullptr);
+  explicit TrackInfoDialog(const Track &track,
+                           std::shared_ptr<Playlist::Playlist> playlist = nullptr,
+                           QWidget *parent = nullptr);
   ~TrackInfoDialog();
+
+signals:
+  void tagEditorOpened(TagEditorDialog *editor);
 
 private slots:
   void on_copy(const QPoint &pos);
   void on_toolButtonOpenFileManager_clicked();
+  void on_toolButtonEditTags_clicked();
   void on_search(const QPoint &pos);
   void on_labelCoverArt_customContextMenuRequested(const QPoint &pos);
+  void refresh_track(const QList<quint64> &uids);
 
 private:
   Ui::TrackInfoDialog *ui;
 
   QStandardItemModel model;
+  Track _track;
+  std::shared_ptr<Playlist::Playlist> _playlist;
+  QString base_title;
   QString track_path;
   QString cover_art_path;